Remove realworld example.
I'd prefer to have one but I'd rather not have one then one that doesn't compile.

Ideally someone could take up maintaining a realworld example outside of SQLx. Until GATs though we
should keep this to one database. Looking back, most of the reason this is so hard to migrate to 0.4 is
because of how generic the bounds are.